requirements.txt
# Core scientific stack
numpy>=1.24
scipy>=1.10
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12

# RL training
stable-baselines3>=2.0.0
gymnasium>=0.29
torch>=2.0

# Configuration and serialisation
pyyaml>=6.0
jsonschema>=4.17

# Testing
pytest>=7.4
pytest-cov>=4.1

# Statistical analysis
statsmodels>=0.14

# ROS2 dependencies are managed through colcon/rosdep, NOT through pip.
# They are listed here for documentation only:
#   - rclpy
#   - std_msgs
#   - geometry_msgs
#   - sensor_msgs
#   - rosbag2_py

# Notebook (optional, for analysis)
jupyter>=1.0
nbformat>=5.9
